+++ src_cpp/src/Engine/game/map/Chunk.h
@@ -76,9 +76,11 @@ public:
     }
 
     //------- get -------//
+    /*
     inline const glm::vec2 get_fpos() const {
         return mcpos.get_fpos();
     }
+    */
     inline const IntVec2 &get_mpos() const {
         return this->mcpos.get_mpos();
     }
@@ -131,7 +133,7 @@ private:
     std::vector<fieldKey_t> fieldKeys {}; //- 8*8 fieldKeys,仅用来辅助 chunk创建
                                           
 
-    float  zOff {}; //- chunk间存在前后层次，
+    double  zOff {}; //- chunk间存在前后层次，
 };
